requirements.txt
streamlit
biopython
py3Dmol
requests
pandas
